requirements.txt
discord.py==2.6.4
yt-dlp==2025.11.12
PyNaCl==1.5.0
imageio-ffmpeg==0.6.0